Makes the PortfolioSection use the WrappingCardsSection to wrap 3 instances of a new PortfolioItemCard widget
This PortfolioItemCard widget currently just fixes itself to a 16/9 aspect ratio, rounds its borders, and provides a drop shadow.
When implemented, the PortfolioItemCard will contain a preview for each item in the portfolio.
When tapped, the item cards will open a modal dialog that will contain more information about the portfolio item.