Reworks the LandingPage navigation bar to control the scroll view, instead of different routes
The scroll_to_index package is added as a dependency.
Using this package, the LandingPage will now scroll to a certain section when a navigation button is pressed, as well as update the selected navigation button when the scroll view is scrolled.

The LandingPage keeps track of the offsets of each section inside the main column.
When the scroll view is scrolled, its offset is compared to that of each section in the main column.
The section with the lowest absolute difference between its offset in the column, and the scroll view offset, is considered the currently selected one.