Adjusts the sections of the landing page so that when they are scrolled to using the navigation buttons, their content appears below the AppBar
Now that the Scaffold body extends behind the app bar for the image in the WelcomeSection, each section must apply a padding at its top with a height equal to that of the AppBar.
The separatorBuilder is removed from the LandingPage's ListView.

Additionally, the section headers have been made to be left-aligned, as specified in the UI design