Makes the contact section have a height at least 80% of the screen height
This is so that if the user scrolls to it, the navigation bar can update to have the contact section selected.

On large screens, it could not be selected due to it being the final item in the list, and having a height not large enough to make it the closest section to the top of the screen.